requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.28.0,<0.29
Pillow>=8.2.0,<8.3.0
requests>=2.26.0,<3.0
pyAudioAnalysis>=0.3.14,<0.4
eyeD3>=0.9.7,<1.0.0
numpy>=1.21.0,<1.26.0
pydub>=0.25.1,<0.26.0
scipy>=1.7.0,<1.16.0
matplotlib>=3.4.2,<3.9.0
tqdm>=4.64.0,<4.66.0
scikit-learn>=1.6.1,<1.7.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1